+++ src/extension/errors/errorHandler.ts
