+++ src/extension/errors/errorRendererComms.ts
