+++ src/extension/errors/interactiveCellResultError.ts
