+++ src/extension/errors/jupyterConnectError.ts
