+++ src/extension/errors/jupyterDebuggerRemoteNotSupportedError.ts
