+++ src/extension/errors/jupyterInstallError.ts
