+++ src/extension/errors/jupyterInvalidKernelError.ts
