+++ src/extension/errors/jupyterKernelDependencyError.ts
