+++ src/extension/errors/jupyterSelfCertsError.ts
