+++ src/extension/errors/jupyterWaitForIdleError.ts
