+++ src/extension/errors/kernelConnectionTimeoutError.ts
