+++ src/extension/errors/kernelDiedError.ts
