+++ src/extension/errors/kernelInterruptTimeoutError.ts
