+++ src/extension/errors/kernelProcessExitedError.ts
