+++ src/extension/errors/sessionDisposedError.ts
