+++ src/intellisense/diagnosticsProvider.ts
