+++ src/intellisense/fileBasedCancellationStrategy.ts
