+++ src/intellisense/intellisenseProvider.ts
