+++ src/intellisense/logReplayService.ts
