+++ src/intellisense/pythonKernelCompletionProvider.ts
