+++ src/notebooks/controllers/fontAwesomeLoader.js
