+++ src/notebooks/helpers.ts
@@ -822,3 +822,171 @@ export function findAssociatedNotebookDocument(
     }
     return notebook;
 }
+
+// eslint-disable-next-line complexity
+export function updateNotebookMetadata(
+    metadata?: nbformat.INotebookMetadata,
+    kernelConnection?: KernelConnectionMetadata,
+    kernelInfo?: Partial<KernelMessage.IInfoReplyMsg['content']>
+) {
+    let changed = false;
+    let kernelId: string | undefined;
+    if (!metadata) {
+        return { changed, kernelId };
+    }
+
+    // If language isn't specified in the metadata, ensure we have that.
+    if (!metadata?.language_info?.name) {
+        metadata = metadata || <nbformat.INotebookMetadata>{ orig_nbformat: 3, language_info: { name: '' } };
+        metadata.language_info = metadata.language_info || { name: '' };
+    }
+
+    let language: string | undefined;
+    switch (kernelConnection?.kind) {
+        case 'connectToLiveKernel':
+            language = kernelConnection.kernelModel.language;
+            break;
+        case 'startUsingRemoteKernelSpec':
+        case 'startUsingLocalKernelSpec':
+            language = kernelConnection.kernelSpec.language;
+            break;
+        case 'startUsingPythonInterpreter':
+            language = PYTHON_LANGUAGE;
+            break;
+        default:
+            break;
+    }
+    if (metadata.language_info.name !== language && language) {
+        metadata.language_info.name = language;
+        changed = true;
+    }
+
+    if (kernelInfo && 'language_info' in kernelInfo && kernelInfo.language_info) {
+        if (!fastDeepEqual(metadata.language_info, kernelInfo.language_info)) {
+            metadata.language_info = cloneDeep(kernelInfo.language_info);
+            changed = true;
+        }
+    } else {
+        // Get our kernel_info and language_info from the current notebook
+        const isPythonConnection = isPythonKernelConnection(kernelConnection);
+        const interpreter = isPythonConnection
+            ? getInterpreterFromKernelConnectionMetadata(kernelConnection)
+            : undefined;
+        if (
+            interpreter &&
+            interpreter.version &&
+            metadata &&
+            metadata.language_info &&
+            metadata.language_info.version !== interpreter.version.raw
+        ) {
+            metadata.language_info.version = interpreter.version.raw;
+            changed = true;
+        } else if (!interpreter && metadata?.language_info && isPythonConnection) {
+            // It's possible, such as with raw kernel and a default kernelspec to not have interpreter info
+            // for this case clear out old invalid language_info entries as they are related to the previous execution
+            // However we should clear previous language info only if language is python, else just leave it as is.
+            metadata.language_info = undefined;
+            changed = true;
+        }
+    }
+
+    const kernelSpecOrModel =
+        kernelConnection && kernelConnectionMetadataHasKernelModel(kernelConnection)
+            ? kernelConnection.kernelModel
+            : kernelConnection?.kernelSpec;
+    if (kernelConnection?.kind === 'startUsingPythonInterpreter') {
+        // Store interpreter name, we expect the kernel finder will find the corresponding interpreter based on this name.
+        const kernelSpec = kernelConnection.kernelSpec;
+        let displayName = metadata.kernelspec?.display_name || kernelConnection.interpreter.displayName || '';
+
+        // If kernel spec is generated by us, then always store `python3` in the notebook (so that its portable).
+        // However if we've created a kernelspec that points to a real kernelspec in a Python environment, then use the name of that spec.
+        let name = '';
+        switch (getKernelRegistrationInfo(kernelSpec)) {
+            case 'registeredByOldVersionOfExt':
+            case 'registeredByNewVersionOfExt':
+                name = 'python3';
+                break;
+            case 'registeredByNewVersionOfExtForCustomKernelSpec': {
+                const originalNameFromOriginalSpecFile = kernelSpec.metadata?.vscode?.originalSpecFile
+                    ? path.basename(path.dirname(kernelSpec.metadata?.vscode?.originalSpecFile))
+                    : undefined;
+
+                name = originalNameFromOriginalSpecFile || kernelSpec.name;
+                displayName = kernelSpec.metadata?.vscode?.originalDisplayName || displayName;
+                break;
+            }
+            default:
+                name = kernelSpec.name;
+                break;
+        }
+
+        if (metadata.kernelspec?.name !== name) {
+            changed = true;
+            metadata.kernelspec = {
+                name,
+                language: PYTHON_LANGUAGE,
+                display_name: displayName
+            };
+            if (getKernelRegistrationInfo(kernelSpec)) {
+                // For kernels generated by us, store the interpreter information.
+                // We'll leave the kernel spec name as `python3`, this way it will work even in Jupyter or the like.
+                // Else if user opens a notebook that works in jupter,
+                // then in ours they cannot go back to jupyter as `python<hash>` is not necessarily a valid kernel in jupter.
+                metadata.interpreter = {
+                    hash: getInterpreterHash(kernelConnection.interpreter)
+                };
+            }
+        }
+    } else if (kernelSpecOrModel && !metadata.kernelspec) {
+        const originalNameFromOriginalSpecFile = kernelSpecOrModel.metadata?.vscode?.originalSpecFile
+            ? path.basename(path.dirname(kernelSpecOrModel.metadata?.vscode?.originalSpecFile))
+            : undefined;
+        // Add a new spec in this case
+        metadata.kernelspec = {
+            name: originalNameFromOriginalSpecFile || kernelSpecOrModel.name || kernelSpecOrModel.display_name || '',
+            display_name:
+                kernelSpecOrModel.metadata?.vscode?.originalDisplayName ||
+                kernelSpecOrModel.display_name ||
+                kernelSpecOrModel.name ||
+                ''
+        };
+        if (kernelSpecOrModel.language) {
+            metadata.kernelspec.language = kernelSpecOrModel.language;
+        }
+        kernelId = kernelSpecOrModel.id;
+        changed = true;
+    } else if (kernelSpecOrModel && metadata.kernelspec) {
+        const originalNameFromOriginalSpecFile = kernelSpecOrModel.metadata?.vscode?.originalSpecFile
+            ? path.basename(path.dirname(kernelSpecOrModel.metadata?.vscode?.originalSpecFile))
+            : undefined;
+        // Spec exists, just update name and display_name
+        const name = originalNameFromOriginalSpecFile || kernelSpecOrModel.name || kernelSpecOrModel.display_name || '';
+        const displayName =
+            kernelSpecOrModel.metadata?.vscode?.originalDisplayName ||
+            kernelSpecOrModel.display_name ||
+            kernelSpecOrModel.name ||
+            '';
+        const language = kernelSpecOrModel.language || kernelSpecOrModel.language || '';
+        if (
+            metadata.kernelspec.name !== name ||
+            metadata.kernelspec.language !== language ||
+            metadata.kernelspec.display_name !== displayName ||
+            kernelId !== kernelSpecOrModel.id
+        ) {
+            changed = true;
+            metadata.kernelspec.name = name;
+            metadata.kernelspec.display_name = displayName;
+            metadata.kernelspec.language = language;
+            kernelId = kernelSpecOrModel.id;
+        }
+        try {
+            // This is set only for when we select an interpreter.
+            // eslint-disable-next-line @typescript-eslint/no-explicit-any
+            delete (metadata.kernelspec as any).metadata;
+        } catch {
+            // Noop.
+        }
+    }
+    return { changed, kernelId };
+}
