+++ src/notebooks/outputs/plotViewHandler.ts
