+++ src/notebooks/serviceRegistry.ts
